resize Lib Int->CGFloat

+++ ZeroCafePlus-iOS/ZeroCafePlus-iOS/ZeroCafePlus-iOS/CommonFunction.swift
@@ -11,7 +11,7 @@ import UIKit
 class CommonFunction: AnyObject {
     
     // imageName : "pngの名前", w or h :サイズ
-    func resizingImage(imageName imageName :String, w:Int, h:Int) ->UIImage
+    func resizingImage(imageName imageName :String, w:CGFloat, h:CGFloat) ->UIImage
     {
         let image = UIImage(named: imageName)
         let size = CGSize(width: w, height: h)
